more refactoring for new feature
